attached_assets/Pasted-react-dom-development-js-28478-Uncaught-Error-Element-type-is-invalid-expected-a-string-for-built-1747059752669.txt

react-dom.development.js:28478 Uncaught Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ProtectedRoute`.
    at createFiberFromTypeAndProps (react-dom.development.js:28478:17)
    at createFiberFromElement (react-dom.development.js:28504:15)
    at reconcileSingleElement (react-dom.development.js:13986:23)
    at reconcileChildFibers2 (react-dom.development.js:14044:35)
    at reconcileChildren (react-dom.development.js:19186:28)
    at mountIndeterminateComponent (react-dom.development.js:20186:5)
    at beginWork (react-dom.development.js:21626:16)
    at HTMLUnknownElement.callCallback2 (react-dom.development.js:4164:14)
    at Object.invokeGuardedCallbackDev (react-dom.development.js:4213:16)
    at invokeGuardedCallback (react-dom.development.js:4277:31)
react-dom.development.js:18704 The above error occurred in the <ProtectedRoute> component:

    at ProtectedRoute (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/src/lib/protected-route.tsx:20:45)
    at Route (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/@fs/home/runner/workspace/node_modules/.vite/deps/wouter.js?v=21d01bc0:338:16)
    at Switch (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/@fs/home/runner/workspace/node_modules/.vite/deps/wouter.js?v=21d01bc0:394:17)
    at main
    at div
    at div
    at AppShell (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/src/components/AppShell.tsx:96:21)
    at Provider (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/@fs/home/runner/workspace/node_modules/.vite/deps/chunk-OXZDJRWN.js?v=21d01bc0:38:15)
    at TooltipProvider (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/@fs/home/runner/workspace/node_modules/.vite/deps/@radix-ui_react-tooltip.js?v=21d01bc0:63:5)
    at AuthProvider (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/src/hooks/use-auth.tsx:27:32)
    at QueryClientProvider (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/@fs/home/runner/workspace/node_modules/.vite/deps/@tanstack_react-query.js?v=21d01bc0:2805:3)
    at App
    at ThemeProvider (https://b4205e83-1543-41d2-8d5b-a4c76af59825-00-naxpodzlazki.sisko.replit.dev/src/components/ThemeProvider.tsx:25:3)

Consider adding an error boundary to your tree to customize error handling behavior.
Visit https://reactjs.org/link/error-boundaries to learn more about error boundaries.
logCapturedError @ react-dom.development.js:18704
react-dom.development.js:12056 Uncaught Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ProtectedRoute`.
    at createFiberFromTypeAndProps (react-dom.development.js:28478:17)
    at createFiberFromElement (react-dom.development.js:28504:15)
    at reconcileSingleElement (react-dom.development.js:13986:23)
    at reconcileChildFibers2 (react-dom.development.js:14044:35)
    at reconcileChildren (react-dom.development.js:19186:28)
    at mountIndeterminateComponent (react-dom.development.js:20186:5)
    at beginWork (react-dom.development.js:21626:16)
    at beginWork$1 (react-dom.development.js:27465:14)
    at performUnitOfWork (react-dom.development.js:26596:12)
    at workLoopSync (react-dom.development.js:26505:5)
